Abstract feedback submission to its own file
