+++ src/services/index.js
@@ -1,0 +1,3 @@
+import sendFeedback from './sendFeedback';
+
+export { sendFeedback };
